% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_function.R
\name{e_recode}
\alias{e_recode}
\title{A recode-function that updates the variable by reference. Requires data.table.}
\usage{
e_recode(dt, variable, old, new)
}
\arguments{
\item{dt}{Working data.table}

\item{variable}{The variable to recode}

\item{old}{a vector of the old values}

\item{new}{a vector of the new values}
}
\value{
Recoded data.table and prints the changes
}
\description{
A recode-function that updates the variable by reference. Requires data.table.
}
\examples{
e_recode()
}
\keyword{lookup}
\keyword{recode,}
