\name{etools-package}
\alias{etools-package}
\alias{etools}
\docType{package}
\title{
\packageTitle{etools}
}
\description{
\packageDescription{etools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{etools}
\packageIndices{etools}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{etools}

Maintainer: \packageMaintainer{etools}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
