% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_function.R
\name{tab}
\alias{tab}
\title{Tab(ulate), similar to Stata (orders by x, or x,y). Requires data.table.
Data argument defaults to dt, for easy usage tab("x").}
\usage{
tab(x = NULL, y = NULL, df = dt)
}
\arguments{
\item{x}{Variable to be tabulated}

\item{y}{Optional second variable}

\item{df}{the dataframe}
}
\value{
Table of (N) cases ordered by x or x, y.
}
\description{
Tab(ulate), similar to Stata (orders by x, or x,y). Requires data.table.
Data argument defaults to dt, for easy usage tab("x").
}
\examples{
tab()
}
\keyword{tab,}
\keyword{table,}
\keyword{tabulate}
